+++ client/cat3/src/partials/sections/dashboard/workzone/blueprint/popups/dockerParamsCtrl.js
@@ -59,7 +59,7 @@
 						$scope.dockerParams.port = subparam[1];
 					} else if (subparam.indexOf('-link') === 0) {
 						$scope.dockerParams.link = subparam[1];
-					} else if (subparam.indexOf('-name') === 0 || subparam.indexOf('-name') === 0) {
+					} else if (subparam.indexOf('-name') === 0 || subparam.indexOf('--name') === 0) {
 						$scope.dockerParams.name = subparam[1];
 					} else if (subparam.indexOf('v') === 0) {
 						$scope.dockerParams.volumes = subparam[1];
@@ -84,10 +84,10 @@
 						dockerParameters += ' -v ' + $scope.dockerParams.volumes;
 					}
 					if ($scope.dockerParams.volumesFrom) {
-						dockerParameters += ' -volumes-from ' + $scope.dockerParams.volumesFrom;
+						dockerParameters += ' --volumes-from ' + $scope.dockerParams.volumesFrom;
 					}
 					if ($scope.dockerParams.link) {
-						dockerParameters += ' -link ' + $scope.dockerParams.link;
+						dockerParameters += ' --link ' + $scope.dockerParams.link;
 					}
 					if ($scope.dockerParams.environment) {
 						dockerParameters += ' -e ' + $scope.dockerParams.environment;
